requirements.txt
python-telegram-bot==13.13
opencv-python==4.8.0.76
python-dotenv==1.0.0